+++ apps/badgeuser/serializers_v2.py
@@ -143,11 +143,3 @@ class AccessTokenSerializerV2(DetailSerializerV2):
     class Meta:
         apispec_definition = ('AccessToken', {})
 
-    def pre_envelope(self, representation, instance):
-        if self.context.get('include_token', False):
-            representation['access_token'] = instance.token
-        return representation
-
-
-
-
