+++ apps/badgeuser/tests.py
@@ -60,40 +60,6 @@ class AuthTokenTests(BadgrTestCase):
         self.assertEqual(user.cached_token(), second_response.data.get('token'))
         self.assertEqual(Token.objects.get(user=user).key, user.cached_token())
 
-    def test_can_encrypt_decrypt_authcode(self):
-        payload = "fakeentityid"
-        code = encrypt_authcode(payload)
-        decrypted_payload = decrypt_authcode(code)
-        self.assertEqual(payload, decrypted_payload)
-
-    def test_can_use_authcode_exchange(self):
-        user = self.setup_user(authenticate=True)
-        application = Application.objects.create(
-            client_id='testing-authcode',
-            client_type=Application.CLIENT_PUBLIC,
-            authorization_grant_type=Application.GRANT_PASSWORD
-        )
-        ApplicationInfo.objects.create(application=application)
-        accesstoken = BadgrAccessToken.objects.generate_new_token_for_user(user, application, scope='r:profile')
-
-        # can exchange valid authcode for accesstoken
-        authcode = authcode_for_accesstoken(accesstoken)
-        response = self.client.post(reverse('v2_api_auth_code_exchange'), dict(code=authcode))
-        self.assertEqual(response.status_code, 200)
-        results = response.data.get('result', [])
-
-        self.assertEqual(len(results), 1)
-        self.assertDictContainsSubset({'access_token': accesstoken.token}, results[0])
-
-        # cant exchange invalid authcode
-        response = self.client.post(reverse('v2_api_auth_code_exchange'), dict(code="InvalidAuthCode"))
-        self.assertEqual(response.status_code, 400)
-
-        # cant exchange expired authcode
-        expired_authcode = authcode_for_accesstoken(accesstoken, expires_seconds=0)
-        response = self.client.post(reverse('v2_api_auth_code_exchange'), dict(code=expired_authcode))
-        self.assertEqual(response.status_code, 400)
-
 
 class UserCreateTests(BadgrTestCase):
