+++ apps/badgeuser/v2_api_urls.py
@@ -4,7 +4,7 @@ from __future__ import unicode_literals
 from django.conf.urls import url
 
 from badgeuser.api import BadgeUserToken, BadgeUserForgotPassword, BadgeUserEmailConfirm, BadgeUserDetail, \
-    AccessTokenList, AccessTokenDetail, AuthCodeExchange
+    AccessTokenList, AccessTokenDetail
 
 urlpatterns = [
 
@@ -14,9 +14,6 @@ urlpatterns = [
 
     url(r'^auth/tokens$', AccessTokenList.as_view(), name='v2_api_access_token_list'),
     url(r'^auth/tokens/(?P<entity_id>[^/]+)$', AccessTokenDetail.as_view(), name='v2_api_access_token_detail'),
-
-
-    url(r'^auth/code$', AuthCodeExchange.as_view(), name='v2_api_auth_code_exchange'),
 
     url(r'^users/(?P<entity_id>self)$', BadgeUserDetail.as_view(), name='v2_api_user_self'),
     url(r'^users/(?P<entity_id>[^/]+)$', BadgeUserDetail.as_view(), name='v2_api_user_detail'),
