+++ apps/mainsite/urls.py
@@ -3,7 +3,7 @@ from django.conf import settings
 from django.conf.urls import include, url
 
 from mainsite.admin import badgr_admin
-from mainsite.oauth2_api import AuthorizationApiView, TokenView
+from mainsite.oauth2_api import AuthorizationApiView, TokenView, AuthCodeExchange
 
 badgr_admin.autodiscover()
 # make sure that any view/model/form imports occur AFTER admin.autodiscover
@@ -29,6 +29,7 @@ urlpatterns = [
     # OAuth2 provider URLs
     url(r'^o/authorize/?$', AuthorizationApiView.as_view(), name='oauth2_api_authorize'),
     url(r'^o/token/?$', TokenView.as_view(), name='oauth2_provider_token'),
+    url(r'^o/code/?$', AuthCodeExchange.as_view(), name='oauth2_code_exchange'),
     url(r'^o/', include('oauth2_provider.urls', namespace='oauth2_provider')),
 
     # Home
